git-svn-id: svn://svn.sphinxsearch.com/sphinx/trunk@3026 406a0c4d-033a-0410-8de8-e80135713968

+++ src/sphinxstd.h
@@ -794,7 +794,7 @@ public:
 	/// resize
 	void Resize ( int iNewLength )
 	{
-		if ( iNewLength>=m_iLength )
+		if ( (unsigned int)iNewLength>=(unsigned int)m_iLength )
 			Reserve ( iNewLength );
 		m_iLength = iNewLength;
 	}
@@ -2126,7 +2126,8 @@ private:
 	static const int iELEMBITS = sizeof(DWORD) * 8;
 	static const int iBYTESIZE = iTOTALBITS / 8;
 	static const int IELEMENTS = iTOTALBITS / iELEMBITS;
-	static const DWORD uALLBITS = ~(0UL);
+	static const DWORD uALLBITS = (DWORD)(~(0UL));
+
 	STATIC_ASSERT ( IELEMENTS>=1, 8_BITS_MINIMAL_SIZE_OF_VECTOR );
 
 public:
